+++ src/Molecules/AccordionItem/AccordionItem.types.ts
@@ -5,10 +5,11 @@ type Props = {
   /** Setting this prop makes the component controlled */
   expanded?: boolean;
   expandedInitial?: boolean;
-  title: string;
+  title: string | React.ReactNode;
   onClick?: React.MouseEventHandler;
   onToggle?: (expanded: boolean) => void;
   ref?: React.Ref<HTMLButtonElement>;
+  withChevron?: boolean;
 };
 
 export type AccordionItemComponent = React.FC<Props>;