+++ src/Molecules/Input/Select/lib/SingleSelectList/SingleSelectList.tsx
@@ -121,6 +121,7 @@ const StyledOption = styled(Typography)<Partial<OptionProps>>`
   padding-bottom: ${(p) => p.theme.spacing.unit(1)}px;
   color: ${(p) => (p.selected ? p.theme.color.cta : p.theme.color.text)};
   height: ${(p) => p.theme.spacing.unit(6)}px;
+  outline: none;
   white-space: nowrap;
   ${(p) =>
     p.focused